requirements.txt
flask>=3.0
requests>=2.31
pytest>=8.0
google-generativeai>=0.8
chromadb>=0.5
beautifulsoup4>=4.12
